Change UserTrackingClient.requestAuthorization input from UInt64 to Double

+++ Sources/UserTracking/Live.swift
@@ -78,7 +78,7 @@ final actor UserTrackingImpl {
   }
 
   func requestAuthorization(
-    delayFor interval: UInt64
+    delayFor interval: Double
   ) async -> AuthorizationStatus {
     logger.info("request authorization")
